backend/requirements.txt
Flask
flasgger
Flask-Cors
Flask-Bootstrap4
Flask-FontAwesome
redis
pymongo